Refactoring the learningOrchestra code and creating metadata file before async processing
