+++ microservices/database_api_image/database.py
@@ -1,22 +1,14 @@
-from pymongo import MongoClient, errors, ASCENDING
+from pymongo import errors
 from bson.json_util import dumps
 import requests
-from contextlib import closing
-import csv
 import json
-import codecs
-from concurrent.futures import ThreadPoolExecutor
-from queue import Queue
-from datetime import datetime
-import pytz
-
-ROW_ID = "_id"
-METADATA_ROW_ID = 0
 
 
-class DatabaseApi:
+class Dataset:
     MESSAGE_INVALID_URL = "invalid url"
     MESSAGE_DUPLICATE_FILE = "duplicate file"
+    ROW_ID = "_id"
+    METADATA_ROW_ID = 0
 
     def __init__(self, database_object, file_manager_object):
         self.database_object = database_object
@@ -24,8 +16,8 @@ class DatabaseApi:
 
     def add_file(self, url, filename):
         try:
-            self.file_manager_object.storage_file(filename, url,
-                                                  self.database_object)
+            self.file_manager_object.save_file(filename, url,
+                                               self.database_object)
 
         except requests.exceptions.RequestException:
             raise Exception(self.MESSAGE_INVALID_URL)
@@ -55,139 +47,11 @@ class DatabaseApi:
 
         for file in self.database_object.get_filenames():
             metadata_file = self.database_object.find_one_in_file(
-                file, {ROW_ID: METADATA_ROW_ID, "type": file_type}
+                file, {self.ROW_ID: self.METADATA_ROW_ID, "type": file_type}
             )
             if metadata_file is None:
                 continue
-            metadata_file.pop(ROW_ID)
+            metadata_file.pop(self.ROW_ID)
             result.append(metadata_file)
 
         return result
-
-
-class MongoOperations:
-    DATABASE_URL = "DATABASE_URL"
-    DATABASE_PORT = "DATABASE_PORT"
-
-    def __init__(self, database_url, replica_set, database_port, database_name):
-        self.mongo_client = MongoClient(
-            database_url + '/?replicaSet=' + replica_set, int(database_port))
-        self.database = self.mongo_client[database_name]
-
-    def connection(self, filename):
-        return self.database[filename]
-
-    def find_in_file(self, filename, query, skip=0, limit=1):
-        file_collection = self.database[filename]
-        return (
-            file_collection.find(query).sort(ROW_ID, ASCENDING).skip(
-                skip).limit(limit)
-        )
-
-    def delete_file(self, filename):
-        file_collection = self.database[filename]
-        file_collection.drop()
-
-    def get_filenames(self):
-        return self.database.list_collection_names()
-
-    def insert_one_in_file(self, filename, json_object):
-        file_collection = self.database[filename]
-        file_collection.insert_one(json_object)
-
-    def update_one_in_file(self, filename, new_value, query):
-        file_collection = self.database[filename]
-        file_collection.update_one(new_value, query)
-
-    def find_one_in_file(self, filename, query):
-        file_collection = self.database[filename]
-        return file_collection.find_one(query)
-
-
-class CsvDownloader:
-    MAX_QUEUE_SIZE = 1000
-    MAX_NUMBER_THREADS = 3
-    FINISHED = "finished"
-    file_headers = None
-
-    def __init__(self):
-        self.thread_pool = ThreadPoolExecutor(
-            max_workers=self.MAX_NUMBER_THREADS)
-        self.download_treatment_queue = Queue(maxsize=self.MAX_QUEUE_SIZE)
-        self.treatment_save_queue = Queue(maxsize=self.MAX_QUEUE_SIZE)
-
-    def download_file(self, url):
-        with closing(requests.get(url, stream=True)) as r:
-            reader = csv.reader(
-                codecs.iterdecode(r.iter_lines(), encoding="utf-8"),
-                delimiter=",",
-                quotechar='"',
-            )
-            self.file_headers = next(reader)
-            for row in reader:
-                self.download_treatment_queue.put(row)
-        self.download_treatment_queue.put(self.FINISHED)
-
-    def treatment_file(self):
-        row_count = 1
-        while True:
-            downloaded_row = self.download_treatment_queue.get()
-            if downloaded_row == self.FINISHED:
-                break
-            json_object = {
-                self.file_headers[index]: downloaded_row[index]
-                for index in range(len(self.file_headers))
-            }
-            json_object[ROW_ID] = row_count
-            self.treatment_save_queue.put(json_object)
-            row_count += 1
-        self.treatment_save_queue.put(self.FINISHED)
-
-    def save_file(self, database_connection, filename):
-        while True:
-            json_object = self.treatment_save_queue.get()
-            if json_object == self.FINISHED:
-                break
-            database_connection.insert_one_in_file(filename, json_object)
-        database_connection.update_one_in_file(
-            filename,
-            {ROW_ID: METADATA_ROW_ID},
-            {"$set": {self.FINISHED: True, "fields": self.file_headers}},
-        )
-
-    @staticmethod
-    def validate_csv_url(url):
-        with closing(requests.get(url, stream=True)) as r:
-            reader = csv.reader(
-                codecs.iterdecode(r.iter_lines(), encoding="utf-8"),
-                delimiter=",",
-                quotechar='"',
-            )
-            first_line = next(reader)
-            first_symbol_html = "<"
-            first_symbol_json = "{"
-            if (
-                    first_line[0][0] == first_symbol_html
-                    or first_line[0][0] == first_symbol_json
-            ):
-                raise requests.exceptions.RequestException
-
-    def storage_file(self, filename, url, database_connection):
-        self.validate_csv_url(url)
-
-        timezone_london = pytz.timezone("Etc/Greenwich")
-        london_time = datetime.now(timezone_london)
-
-        metadata_file = {
-            "datasetName": filename,
-            "url": url,
-            "timeCreated": london_time.strftime("%Y-%m-%dT%H:%M:%S-00:00"),
-            ROW_ID: METADATA_ROW_ID,
-            self.FINISHED: False,
-            "fields": "processing",
-            "type": "dataset"
-        }
-        database_connection.insert_one_in_file(filename, metadata_file)
-        self.thread_pool.submit(self.download_file, url)
-        self.thread_pool.submit(self.treatment_file)
-        self.thread_pool.submit(self.save_file, database_connection, filename)
