+++ microservices/histogram_image/server.py
@@ -1,7 +1,7 @@
 from flask import jsonify, Flask, request
 import os
-from histogram import MongoOperations, HistogramRequestValidator, Histogram
-from concurrent.futures import ThreadPoolExecutor
+from histogram import Histogram
+from utils import Database, UserRequest, Metadata
 
 HTTP_STATUS_CODE_SUCCESS_CREATED = 201
 HTTP_STATUS_CODE_NOT_ACCEPTABLE = 406
@@ -30,52 +30,47 @@ MICROSERVICE_URI_GET_PARAMS = "?query={}&limit=10&skip=0"
 
 app = Flask(__name__)
 
-thread_pool = ThreadPoolExecutor()
-
 
 @app.route("/histograms", methods=["POST"])
 def create_histogram():
-    database = MongoOperations(
+    parent_filename = request.json[PARENT_FILENAME_NAME]
+    histogram_filename = request.json[HISTOGRAM_FILENAME_NAME]
+    fields_name = request.json[FIELDS_NAME]
+
+    database = Database(
         os.environ[DATABASE_URL],
         os.environ[DATABASE_REPLICA_SET],
         os.environ[DATABASE_PORT],
         os.environ[DATABASE_NAME],
     )
 
-    request_validator = HistogramRequestValidator(database)
+    request_validator = UserRequest(database)
 
     request_errors = analyse_request_errors(
         request_validator,
-        request.json[PARENT_FILENAME_NAME],
-        request.json[HISTOGRAM_FILENAME_NAME],
-        request.json[FIELDS_NAME])
+        parent_filename,
+        histogram_filename,
+        fields_name)
 
     if request_errors is not None:
         return request_errors
 
-    thread_pool.submit(histogram_async_processing,
-                       database,
-                       request.json[PARENT_FILENAME_NAME],
-                       request.json[HISTOGRAM_FILENAME_NAME],
-                       request.json[FIELDS_NAME])
+    metadata = Metadata(database)
+    histogram = Histogram(database, metadata)
+
+    histogram.create_file(
+        parent_filename,
+        histogram_filename,
+        fields_name,
+    )
 
     return (
         jsonify({
             MESSAGE_RESULT:
                 MICROSERVICE_URI_GET +
-                request.json[HISTOGRAM_FILENAME_NAME] +
+                histogram_filename +
                 MICROSERVICE_URI_GET_PARAMS}),
         HTTP_STATUS_CODE_SUCCESS_CREATED,
-    )
-
-
-def histogram_async_processing(database, parent_filename, histogram_filename,
-                               fields_name):
-    histogram = Histogram(database)
-    histogram.create_histogram(
-        parent_filename,
-        histogram_filename,
-        fields_name,
     )
