+++ microservices/pca_image/pca.py
@@ -1,6 +1,5 @@
 from pyspark.sql import SparkSession
 import os
-from pymongo import MongoClient
 import numpy as np
 from sklearn.decomposition import PCA
 from sklearn.preprocessing import LabelEncoder
@@ -12,13 +11,13 @@ SPARKMASTER_PORT = "SPARKMASTER_PORT"
 SPARK_DRIVER_PORT = "SPARK_DRIVER_PORT"
 PCA_HOST_NAME = "PCA_HOST_NAME"
 IMAGES_PATH = "IMAGES_PATH"
-IMAGE_FORMAT = ".png"
 
 
-class PcaGenerator:
+class Pca:
     MONGO_SPARK_SOURCE = "com.mongodb.spark.sql.DefaultSource"
     DOCUMENT_ID = "_id"
     METADATA_FILE_ID = 0
+    IMAGE_FORMAT = ".png"
 
     def __init__(self, database_url_input):
         self.spark_session = (
@@ -55,7 +54,8 @@ class PcaGenerator:
         treated_array = np.array(encoded_dataframe)
         embedded_array = PCA(n_components=2).fit_transform(treated_array)
         embedded_array = pandas.DataFrame(embedded_array)
-        image_path = os.environ[IMAGES_PATH] + "/" + pca_filename + IMAGE_FORMAT
+        image_path = os.environ[
+                         IMAGES_PATH] + "/" + pca_filename + self.IMAGE_FORMAT
 
         if label_name is not None:
             embedded_array[label_name] = encoded_dataframe[label_name]
@@ -104,85 +104,3 @@ class PcaGenerator:
                     text_fields.append(column)
 
         return text_fields
-
-
-class MongoOperations:
-    def __init__(self, database_url, replica_set, database_port, database_name):
-        self.mongo_client = MongoClient(
-            database_url + '/?replicaSet=' + replica_set, int(database_port))
-        self.database = self.mongo_client[database_name]
-
-    def find_one(self, filename, query):
-        file_collection = self.database[filename]
-        return file_collection.find_one(query)
-
-    def get_filenames(self):
-        return self.database.list_collection_names()
-
-    @staticmethod
-    def collection_database_url(database_url, database_name, database_filename,
-                                database_replica_set
-                                ):
-        return (
-                database_url
-                + "/"
-                + database_name
-                + "."
-                + database_filename
-                + "?replicaSet="
-                + database_replica_set
-                + "&authSource=admin"
-        )
-
-
-class PcaRequestValidator:
-    MESSAGE_INVALID_FILENAME = "invalid dataset name"
-    MESSAGE_DUPLICATE_FILE = "duplicate file"
-    MESSAGE_INVALID_LABEL = "invalid field"
-    MESSAGE_NOT_FOUND = "file not found"
-    MESSAGE_UNFINISHED_PROCESSING = "unfinished processing in input dataset"
-
-    def __init__(self, database_connector):
-        self.database = database_connector
-
-    def parent_filename_validator(self, filename):
-        filenames = self.database.get_filenames()
-
-        if filename not in filenames:
-            raise Exception(self.MESSAGE_INVALID_FILENAME)
-
-    def finished_processing_validator(self, filename):
-        filename_metadata_query = {"datasetName": filename}
-
-        filename_metadata = self.database.find_one(filename,
-                                                   filename_metadata_query)
-
-        if filename_metadata["datasetName"] == False:
-            raise Exception(self.MESSAGE_UNFINISHED_PROCESSING)
-
-    @staticmethod
-    def pca_filename_existence_validator(pca_filename):
-        images = os.listdir(os.environ[IMAGES_PATH])
-        image_name = pca_filename + IMAGE_FORMAT
-        if image_name in images:
-            raise Exception(PcaRequestValidator.MESSAGE_DUPLICATE_FILE)
-
-    @staticmethod
-    def pca_filename_nonexistence_validator(pca_filename):
-        images = os.listdir(os.environ[IMAGES_PATH])
-        image_name = pca_filename + IMAGE_FORMAT
-
-        if image_name not in images:
-            raise Exception(PcaRequestValidator.MESSAGE_NOT_FOUND)
-
-    def filename_label_validator(self, filename, label):
-        if label is None:
-            return
-
-        filename_metadata_query = {"datasetName": filename}
-
-        filename_metadata = self.database.find_one(filename,
-                                                   filename_metadata_query)
-
-        if label not in filename_metadata["fields"]:
-            raise Exception(self.MESSAGE_INVALID_LABEL)
